requirements.txt
streamlit
python-dotenv
PyPDF2
reportlab
requests